Change test to Junit
